refactor: new Extensions folder, refactor extensions
